Make Ember.Error available in ember-metal.
