+++ packages/container/tests/container_test.js
@@ -14,7 +14,7 @@ var o_create = Object.create || (function(){
 
   return function(o) {
     if (arguments.length !== 1) {
-      throw new Error('Object.create implementation only accepts one parameter.');
+      throw new Ember.Error('Object.create implementation only accepts one parameter.');
     }
     F.prototype = o;
     return new F();
