+++ packages/ember-metal/tests/error_test.js
