+++ packages/ember-metal/tests/run_loop/unwind_test.js
@@ -5,7 +5,7 @@ test('RunLoop unwinds despite unhandled exception', function() {
 
   raises(function() {
     Ember.run(function() {
-      Ember.run.schedule('actions', function() { throw new Error("boom!"); });
+      Ember.run.schedule('actions', function() { throw new Ember.Error("boom!"); });
     });
   }, Error, "boom!");
 
@@ -25,9 +25,9 @@ test('Ember.run unwinds despite unhandled exception', function() {
 
   raises(function() {
     Ember.run(function() {
-      throw new Error("boom!");
+      throw new Ember.Error("boom!");
     });
-  }, Error, "boom!");
+  }, Ember.Error, "boom!");
 
   equal(Ember.run.currentRunLoop, initialRunLoop, "Previous run loop should be cleaned up despite exception");
